refactor: rename & simplify
